qdf: move endian + reflect-alloc helpers to internal/, document public API
Cleans up the root package layout per Go best practice — pure
helpers that don't touch the codec's package-private types move
under internal/, and the public API surface gets an explicit
contract block in the qdf.go package doc.

Moves:

  endian_be.go, endian_le.go
      -> internal/endian/endian_be.go, internal/endian/endian_le.go
      with the exported const renamed nativeLittleEndian
      -> endian.NativeIsLittle. qpack.go is the only caller and
      now imports internal/endian directly.

  reflect_alloc.go, reflect_alloc_reflect2.go
      -> internal/reflectutil/alloc.go, internal/reflectutil/alloc_reflect2.go
      with the trio exported as MakeSlice / SliceData / MakeMap.
      reflect_encode.go is the only caller and routes through the
      new package; the qdf_reflect2 build-tag swap stays in place.

Public API surface (already exists today) is now documented in
the package doc on qdf.go: top-level entry points, generic
wrappers, low-level Encoder / Decoder method list, Marshaler /
Unmarshaler hook points. Internal packages are explicitly
called out as "implementation detail, may change between
releases" so downstream importers know where the stability line
sits.

Also includes the two outstanding linter fixes:
  - gofmt re-alignment on state.go mruRing constant block
  - nolint:prealloc on a literal-slice constructor in
    preintern_test.go where the prealloc hint was misleading
